+++ lib/doorkeeper/request/client_credentials.rb
@@ -3,14 +3,14 @@
 module Doorkeeper
   module Request
     class ClientCredentials < Strategy
-      delegate :client, :parameters, to: :server
+      delegate :client, :dpop_proof, :parameters, to: :server
 
       def request
         @request ||= OAuth::ClientCredentialsRequest.new(
           Doorkeeper.config,
           client,
           parameters,
-        )
+        ).tap { |request| request.dpop_proof = dpop_proof }
       end
     end
   end