+++ spec/dummy/db/schema.rb
@@ -2,11 +2,11 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
@@ -26,8 +26,8 @@ ActiveRecord::Schema.define(version: 20260801000000) do
     t.string "tenant_name"
     t.bigint "access_token_id"
     unless ENV["WITHOUT_PKCE"]
-      t.string   "code_challenge"
-      t.string   "code_challenge_method"
+      t.string "code_challenge"
+      t.string "code_challenge_method"
     end
     t.index ["access_token_id"], name: "index_oauth_access_grants_on_access_token_id"
     t.index ["token"], name: "index_oauth_access_grants_on_token", unique: true
@@ -46,6 +46,7 @@ ActiveRecord::Schema.define(version: 20260801000000) do
     t.text "resource"
     t.string "previous_refresh_token", default: "", null: false
     t.string "tenant_name"
+    t.string "dpop_jkt"
     t.index ["refresh_token"], name: "index_oauth_access_tokens_on_refresh_token", unique: true
     t.index ["resource_owner_id"], name: "index_oauth_access_tokens_on_resource_owner_id"
     t.index ["token"], name: "index_oauth_access_tokens_on_token", unique: true
@@ -74,5 +75,4 @@ ActiveRecord::Schema.define(version: 20260801000000) do
     t.datetime "updated_at"
     t.string "password"
   end
-
 end